requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.7.2
autopep8==2.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-rest-framework==0.1.0
django-restframework==0.0.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
drf-yasg2==1.19.4
environ==1.0
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
promise==2.3
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
